requirements.txt
flask>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-mock>=3.12.0
flake8>=6.1.0
